+++ src/main/scala/org/podval/tools/publish/markup/Entity.scala
@@ -1,7 +1,9 @@
 package org.podval.tools.publish.markup
 
-import org.podval.xml.WithRawXml
-import zio.blocks.schema.Modifier
+import org.podval.xml.XmlExtras
+import org.podval.xml.codec.XmlCodec
+import zio.blocks.schema.{Modifier, Schema}
+import zio.blocks.typeid.TypeId
 
 //final class Entity(
 //  val entityType: EntityType,
@@ -10,8 +12,24 @@ import zio.blocks.schema.Modifier
 //  val mainName: String  // Note: can mostly be reconstructed from the name...
 //)
 final case class Entity(
-  @Modifier.config("xml.attribute", "") id: Option[String],
+  @Modifier.config(XmlCodec.Attribute, "") id: Option[String] = None,
 //  val entityType: EntityType,
-  @Modifier.config("xml.attribute", "") role: Option[String],
-  names: Seq[EntityName],
-) extends WithRawXml
+  @Modifier.config(XmlCodec.Attribute, "") role: Option[String] = None,
+  @Modifier.config(XmlCodec.Element, "persName")
+  @Modifier.alias("placeName")
+  @Modifier.alias("orgName")
+  names: Seq[EntityName] = Seq.empty,
+  extras: XmlExtras = XmlExtras()
+) derives CanEqual
+
+object Entity:
+  given schema: Schema[Entity] = Schema.derived
+
+  val codec: XmlCodec[Entity] = XmlCodec.derived
+
+  def codec(kind: EntityKind): XmlCodec[Entity] =
+    schema
+      .deriving(XmlCodec.deriver)
+      .modifier(TypeId.of[Entity], Modifier.config(XmlCodec.Element, kind.element))
+      .modifier(TypeId.of[Entity], "names", Modifier.config(XmlCodec.Element, kind.nameElement))
+      .derive