+++ src/main/scala/org/podval/tools/publish/markup/EntityReference.scala
@@ -1,11 +1,25 @@
 package org.podval.tools.publish.markup
 
-import org.podval.xml.WithRawXml
-import zio.blocks.schema.Modifier
+import org.podval.xml.XmlExtras
+import org.podval.xml.codec.XmlCodec
+import zio.blocks.schema.{Modifier, Schema}
+import zio.blocks.typeid.TypeId
 
 final case class EntityReference(
-//  @Modifier.config("xml.attribute", "") `type`: EntityType,
-  @Modifier.config("xml.attribute", "") id: Option[String],
-  @Modifier.config("xml.attribute", "") role: Option[String],
-  @Modifier.config("xml.attribute", "") ref: Option[String]
-) extends WithRawXml
+//  @Modifier.config(XmlCodec.Attribute, "") `type`: EntityType,
+  @Modifier.config(XmlCodec.Attribute, "") id: Option[String] = None,
+  @Modifier.config(XmlCodec.Attribute, "") role: Option[String] = None,
+  @Modifier.config(XmlCodec.Attribute, "") ref: Option[String] = None,
+  extras: XmlExtras = XmlExtras()
+) derives CanEqual
+
+object EntityReference:
+  given schema: Schema[EntityReference] = Schema.derived
+
+  val codec: XmlCodec[EntityReference] = XmlCodec.derived
+
+  def codec(kind: EntityKind): XmlCodec[EntityReference] =
+    schema
+      .deriving(XmlCodec.deriver)
+      .modifier(TypeId.of[EntityReference], Modifier.config(XmlCodec.Element, kind.nameElement))
+      .derive