+++ src/main/scala/org/podval/tools/publish/markup/TeiMarkup.scala
@@ -6,21 +6,18 @@ import org.podval.xml.{Xml, Xml2Html, XmlAttribute, XmlUtil}
 import org.podval.xml.XmlUtil.*
 import java.io.File
 
-// TODO @xmlAttribute("id") and @xmlNamespace() annotations mentioned in the documentation
-// do not work! Raw @Modifier.config("xml.attribute", "") does...
-// File the issue.
-// TODO working around ZIO Blocks XML bug where attributes are discarded if the element does not have sub-elements;
-// report it!
+// Binding hints for XmlCodec are @Modifier.config(XmlCodec.Attribute, "") etc.
+// Schema.derived ignores @xmlAttribute / @xmlNamespace (they are not Modifier).
 
-//@Modifier.config("xml.namespace.uri", "http://www.tei-c.org/ns/1.0")
-//@Modifier.config("xml.namespace.prefix", "tei")
+//@Modifier.config(XmlCodec.NamespaceUri, "http://www.tei-c.org/ns/1.0")
+//@Modifier.config(XmlCodec.NamespacePrefix, "tei")
 //final case class TEI(
 //  text: Text
 //)
 //
 //final case class Text(
-//  @Modifier.config("xml.attribute", "") lang: Option[String] = None,
-//  body: RawXml
+//  @Modifier.config(XmlCodec.Attribute, "") lang: Option[String] = None,
+//  body: XmlNode.Element
 //)
 object TeiMarkup extends Markup(
   name = "TEI",